requirements.txt
amqp==5.0.6
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.0.9
Django==3.2.5
django-extensions==3.1.3
djangorestframework==3.12.4
flower==1.0.0
humanize==3.10.0
idna==3.2
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
kombu==5.1.0
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.9.0
python-dotenv==0.18.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
tornado==6.1
traitlets==5.0.5
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
